pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codesage"
version = "0.2.1"
description = "CodeSage starter: ingestion, parsing, embeddings (API only)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="CodeSage"}]

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "GitPython>=3.1",
  "SQLAlchemy>=2.0",
  "numpy>=1.24",
]

[project.optional-dependencies]
embeddings = ["transformers>=4.38", "torch>=2.1"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["codesage*"]
